Refactoring FOFTable to remove the use of getProperties()
Ideally I'd want to go with a full JRegistry implementation and use __get / __set / __call to access the table properties, similar to FOFModel. However this is slow and backwards incompatible so I went with a compromise solution. You can now use the setKnownFields / addKnownField methods to let FOFTable know which fields belong to the database table.

Since I was at it, I improved the magic fields handling and added a few database escapes that were forgotten throughout the code (mostly on integer values, but still!).