removed a probably unnecessary div and moved hidden inputs to the end of the form
a) The Joomla3 renderer enclose the output in a div with a class joomla-version-VERSIONNUMBER these div adds a gab between the button bar and the data area. (GAP1:red)
b) When the hidden input fields are at the beginning of the form the div class="span2" is not longer the first child and then the normal left padding isn't set to 0. (GAP2:green)

for details see http://d.pr/i/7DNO