Re-sync submenu for categories view
When you integrate categories in your component you don't have the link bar from your component when you are on a category view like banner for example. The Categories Extension calls a function ExtensionName::addSubmenu so you could do it the old way and add the items here but then you are out of sync with your Submenu definition. This patch add a function renderCategoryLinkbar to the Render that re-sync the toolbar