too much divs closed in J3
This isn't really **viewable** in the backend, but in the frontend when you have left or right column and these flip under the form. It also makes a the HTML invalid, but for the backend it doesn't it makes it more worse, because if you have 30 or 32 errors doesn't count so much.