ServerSide Validation and fix for uncheck checkboxes
re-submit to staging as requested

added server side validation for save, you need to add a attribute in the form tag serverside_validation="true" to validate the data

I have also created a branch (https://github.com/akeeba/fof/tree/feature/serverside-validation) with these changes but it is not so much so I think a pull request can do the job.

The only thing I don't like so much is the attribute in the form definition, I think a variable validation_processor {js, server, js+server} is better extensible but this could make changes in old components necessary depending on what we chose as the default value.