changed var _hasForm to protected to allow changing in a derived class
Having the variable privat makes it hard to use a form in a custom task. I have a custom workflow:

add -> preview -> change -> preview ..... -> save
			|
  			-> save

This is made with two custom task "preview" and "change". Change is almost that same as add but without resetting the model, I can't use edit because my data are only saved in the session.