Closes gh-178 JFormField changes in Joomla! 3.2 Beta 1 broke FOF form fields. BACKWARDS INCOMPATIBLE CHANGE.
Backwards compatibility notes
----------------------------------------
Since Joomla! 3.2 Beta 1 JFormField comes with a magic __set method which drops
property assignments of any property JFormField is not aware of. This means that
in order for your custom fields to work you need to add the following two public
variables in your custom field class:

	/** @var   FOFTable  The item being rendered in a repeatable form field */
	public $item;
	
	/** @var int A monotonically increasing number, denoting the row number in a repeatable view */
	public $rowid;

Failure to include them will result in an error message when trying to use your
custom field in a browse form.