Fixing stupid template developers
Several template developers chose to screw up Joomla! and load NEITHER Bootstrap compatible stylesheets NOR Bootstrap Javascript. This completely breaks extensions which, of course, rely on core Joomla! code not being broken. Our only defense is to force load a FULL namespaced copy of Bootstrap CSS and a full copy of the Bootstrap Javascript only in the front-end and hope for the best...