View: Create a view finder through the factory
This allows varying degrees of magic depending on the factory used. The basic and magic factories will try for an exact view template match. The switch and magic switch factories will look for the singularised/plural view, with or without a subtemplate and with the specified or default layout (i.e. how FOF 2.x worked).