Built-in support for locked records
If you try to run a DataController task which modifies a locked record it will try to check it in first. If it fails you'll get an error and back you go to the browse task.

DataModel has a new checkin() method to check in items. This is different than unlock(); it checks if you should be able to unlock a record before trying to unlock it.

The XML form fields SelectRow and Text now have a checkout attribute. Set it to "true" to add locked report support:

* On SelectRow it makes the checkbox disappear on locked records.
* On Text fields it adds a "Check-in" button before the text, useful for title fields

If you want the Check-In button in the toolbar to work you must not enable SelectRow's checkbox attribute. If you do you will be unable to select the row to check in!