XML Form Text field: support for external formatter
Use the format_source_file, format_source_class and format_source_method to have a static method of another class (optionally, defining the file it is located in) format the value. This eliminates the need for creating custom fields when you just need a conditional format which can't be expressed with a static string. Just create a helper and you're done!