XML Forms: fieldsets can include external view templates
Each fieldset of a Read and Edit form can optionally include an external view template (PHP or Blade) or even an entirely different View of the same or a different model in a fieldset. The external view will appear before the fieldset's fields. If you don't want any fields to appear just create a fieldset without any fields. The following attributes are used to specify an external view.

**source**

(mandatory) The URI of the view template you want to load. For example `admin:com_example/Items/summary`

**source_component**

(optional) The component providing the View object class to render the view template. By default it's the component the XML form lives in. Otherwise specify the component name, e.g. `com_example`.

**source_view**

(optional) The name of the view for the View object class which renders the view template. By default FOF will use a new `FOF30\View\View` object instance with the name `FAKE_FORM_VIEW`. Otherwise specify the view name, e.g. 'Items'.

The view name is fed through the `view()` method of the Factory object of the container for the component specified in `source_component`. This means that if the component uses a magic factory you don't even need to have a concrete View class.

**source_view_type**

(optional) The view type (e.g. `raw`, `html`, `form`, `json`, `csv`, ...) for the View object class which renders the view template. By default FOF will use `html`. The view type is fed through the `view()` method of the Factory object of the container for the component specified in `source_component`.

FOF will create a View object based on the information above. It will then call its `populateFromModel` method, passing it the Model object used to render the XML form. You can override that method on your View class if you need to set up view properties based on the data contained in the model

FOF also passes a few variables to the view template:
* model (type `FOF30\Model\DataModel`) The model of the form being rendered
* form (type `FOF30\Form\Form`) The form currently being rendered
* fieldset (type `SimpleXMLElement`) The fieldset currently being rendered
* formType (type `string`) The type of the form being rendered, e.g. "read" or "edit"
* innerHtml (type `array`) See `FOF30\Render\RenderInterface::renderFieldset` for information