Model Filters behavior: special handling for primary key
If the mode's ignore request flag is set we'll also look for an 'id' state variable if a state variable by the same name as the key doesn't exist. If ignore request is not set in the model we do not filter by 'id' since this interferes with going from an edit page to a browse page (the list is filtered by id without user controls to unset it).